@@ -10,6 +10,7 @@ using Microsoft.AspNetCore.Mvc.Authorization;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json.Serialization;
 using Ranger.Common;
@@ -21,36 +22,36 @@ namespace Ranger.Services.Integrations
 {
     public class Startup
     {
+        private readonly IWebHostEnvironment Environment;
         private readonly IConfiguration configuration;
-        private readonly ILoggerFactory loggerFactory;
-        private readonly ILogger<Startup> logger;
-        private IContainer container;
+        private ILoggerFactory loggerFactory;
         private IBusSubscriber busSubscriber;
 
-        public Startup(IConfiguration configuration, ILoggerFactory loggerFactory, ILogger<Startup> logger)
+        public Startup(IWebHostEnvironment environment, IConfiguration configuration)
         {
+            this.Environment = environment;
             this.configuration = configuration;
-            this.loggerFactory = loggerFactory;
-            this.logger = logger;
         }
 
-        public IServiceProvider ConfigureServices(IServiceCollection services)
+        public void ConfigureServices(IServiceCollection services)
         {
-            services.AddMvcCore(options =>
-            {
-                var policy = ScopePolicy.Create("projectsApi");
-                options.Filters.Add(new AuthorizeFilter(policy));
-                options.ModelBindingMessageProvider.SetValueMustNotBeNullAccessor(
-                    (_) => "The field is required.");
-            })
-                .AddAuthorization()
-                .AddJsonFormatters()
-                .AddJsonOptions(options =>
+            services.AddControllers(options =>
+                {
+                    options.EnableEndpointRouting = false;
+                })
+                .AddNewtonsoftJson(options =>
                 {
                     options.SerializerSettings.ContractResolver = new CamelCasePropertyNamesContractResolver();
                     options.SerializerSettings.ReferenceLoopHandling = Newtonsoft.Json.ReferenceLoopHandling.Ignore;
                 });
 
+            services.AddAuthorization(options =>
+            {
+                options.AddPolicy("integrationsApi", policyBuilder =>
+                    {
+                        policyBuilder.RequireScope("integrationsApi");
+                    });
+            });
             services.AddEntityFrameworkNpgsql().AddDbContext<IntegrationsDbContext>((serviceProvider, options) =>
             {
                 options.UseNpgsql(configuration["cloudSql:ConnectionString"]);
@@ -81,19 +82,20 @@ namespace Ranger.Services.Integrations
             services.AddDataProtection()
                 .ProtectKeysWithCertificate(new X509Certificate2(configuration["DataProtectionCertPath:Path"]))
                 .PersistKeysToDbContext<IntegrationsDbContext>();
+        }
 
-            var builder = new ContainerBuilder();
-            builder.Populate(services);
+        public void ConfigureContainer(ContainerBuilder builder)
+        {
             builder.RegisterInstance<CloudSqlOptions>(configuration.GetOptions<CloudSqlOptions>("cloudSql"));
             builder.RegisterType<IntegrationsDbContext>().InstancePerDependency();
             builder.RegisterAssemblyTypes(typeof(BaseRepository<>).Assembly).AsClosedTypesOf(typeof(BaseRepository<>)).InstancePerDependency();
             builder.AddRabbitMq(loggerFactory);
-            container = builder.Build();
-            return new AutofacServiceProvider(container);
         }
 
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env, IApplicationLifetime applicationLifetime)
+        public void Configure(IApplicationBuilder app, IHostApplicationLifetime applicationLifetime, ILoggerFactory loggerFactory)
         {
+            this.loggerFactory = loggerFactory;
+
             applicationLifetime.ApplicationStopping.Register(OnShutdown);
             app.UseAuthentication();
             app.UseMvcWithDefaultRoute();
